refactor(control-api): extract create/publish from response_actions/mod.rs
